backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.0
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
openpyxl==3.1.5
PyPDF2==3.0.1
aiofiles==24.1.0
httpx==0.28.0
